btrfs: Allow replacing device with a smaller one if possible
As long as the target device has enough capacity for the total bytes
of the source device, allow the replacement to occur.

Just changing the size validation isn't enough though, since the
rest of the replacement code just assumes that the source device is
identical to the target device. The current code just blindly
copies the total disk size from the source to the target.

A btrfs resize <devid>:max could be performed, but we might as well
just set the disk size for the new target device correctly in the
first place before initiating a scrub, which is what this patch does.

When initializing the target device, the size in bytes is calculated
in the same way that btrfs_init_new_device does it.

When the replace operation completes, btrfs_dev_replace_finishing no
longer clobbers the target device size with the source device size.

Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=112741